Support display embeded web link
Refactor PostEmbed component and add new subcomponents  Reorganized PostEmbed component to delegate tasks to newly created subcomponents. Specifically, responsibilities were shifted to PostEmbedRecord, PostEmbedImages, and PostEmbedExternal. This makes the code more modular and easier to maintain since each subcomponent focuses on a specific task.
